move page logic from /containers to /pages
