--- src/containers/user-profile/index.ts
@@ -1,1 +1,0 @@
-export * from "./user-profile.jsx";
