+++ src/styles/pages/leaderboard.ts
