+++ src/styles/pages/player-profile.ts
